fix: added to components types and constants folders
