+++ src/pages/AnalysisResultPage/components/AnswerForUserQuestion/types/index.ts
@@ -1,0 +1,3 @@
+export interface AnswerForUserQuestionProps {
+  questionResponse: { question: string; answer: string };
+}
