+++ src/pages/AnalysisResultPage/components/BloodTestSummary/types/index.ts
@@ -1,0 +1,5 @@
+import type { FullBloodTestAnalysisResult } from '@/pages/AnalysisResultPage/types';
+
+export interface BloodTestSummaryProps {
+  data?: FullBloodTestAnalysisResult;
+}
