+++ src/pages/AnalysisResultPage/components/DeleteMarkerModal/types/index.ts
@@ -1,0 +1,6 @@
+export interface DeleteMarkerModalProps {
+  open: boolean;
+  markerName: string;
+  onClose: () => void;
+  onConfirm: () => void;
+}
