+++ src/pages/AnalysisResultPage/components/HealthBar/index.tsx
@@ -1,7 +1,7 @@
 import { MARKER_STATUS_CLASSES } from '@/constants';
-import type { HealthBarProps } from '@/pages/AnalysisResultPage/types';
 
 import { PositionIndicator, ProgressBar, StretchedIcon } from './styles';
+import type { HealthBarProps } from './types';
 
 const getStatusClass = (status: string): string => {
   return MARKER_STATUS_CLASSES[status as keyof typeof MARKER_STATUS_CLASSES] || '';
