+++ src/pages/AnalysisResultPage/components/HealthBar/types/index.ts
@@ -1,0 +1,5 @@
+export interface HealthBarProps {
+  position: number;
+  status: string;
+  isSmallScreen: boolean;
+}
