+++ src/pages/AnalysisResultPage/components/PersonalizedRecommendations/constants/index.ts
