+++ src/pages/AnalysisResultPage/components/RecommendationCard/types/index.ts
@@ -1,0 +1,8 @@
+export interface RecommendationSectionProps {
+  title: string;
+  icon: React.ElementType;
+  items: string[];
+  bgColor: string;
+  iconColor: string;
+  bgIconColor: string;
+}
