+++ src/pages/AnalysisResultPage/components/RecommendationsList/types/index.ts
@@ -1,0 +1,8 @@
+import type { RecommendationDescriptions } from '@/pages/AnalysisResultPage/types';
+
+export interface RecommendationsProps {
+  supplementsRecommendations?: RecommendationDescriptions;
+  nutritionRecommendations?: RecommendationDescriptions;
+  drugsRecommendations?: RecommendationDescriptions;
+  exerciseRecommendations?: RecommendationDescriptions;
+}
