+++ src/pages/AnalysisResultPage/components/ResultButtons/constants/index.ts
@@ -1,0 +1,8 @@
+export const PDF_POLL_INTERVAL = 2000;
+export const ICON_SIZE_SMALL = 20;
+
+export const PDF_STATUS = {
+  PENDING: 'pending',
+  COMPLETED: 'completed',
+  FAILED: 'failed',
+} as const;
