+++ src/pages/LoginPage/components/LoginForm/types/index.ts
@@ -1,0 +1,4 @@
+export type LoginFormInputs = {
+  email: string;
+  message: string;
+};
