+++ src/pages/LoginPage/components/SocialLoginButton/types/index.ts
@@ -1,0 +1,6 @@
+export type SocialLoginButtonProps = {
+  icon: React.ReactNode;
+  label: string;
+  onClick?: () => void;
+  linkPath?: string;
+};
