+++ src/pages/UploadPage/components/ProgressIndicator/types/index.ts
@@ -1,0 +1,5 @@
+export interface HealthReportAnalyzerProps {
+  isLoading: boolean;
+  titleKey?: string;
+  descriptionKey?: string;
+}
