+++ src/pages/UploadPage/components/UploadErrorState/types/index.ts
@@ -1,0 +1,3 @@
+export interface UploadErrorStateProps {
+  onRetry: () => void;
+}
