+++ src/pages/UploadPage/components/UploadStep/types/index.ts
@@ -1,0 +1,3 @@
+export interface UploadStepProps {
+  onContinue: () => void;
+}
